refactor: move Visited under iter.rs
